+++ zcash-sign/src/lib.rs
@@ -1,6 +1,7 @@
 mod generate;
 mod sign;
+pub(crate) mod sign_ywallet;
 pub mod transaction_plan;
 
 pub use generate::generate;
-pub use sign::sign;
+pub use sign::{sign, Input};
